opt/bc: merge directly used imm into use site
This merges patterns like

    LoadI { dst, 0 }
    I<cmp> { dst, rhs, .. }

Into

    Nop
    I<cmp> { dst, rhs, .. }

Where
    LoadI.dst == I<cmp>.rhs
    <cmp> := Eq, Gt, Lt
